# app/src/test/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: ""
  description: Generated server url
paths:
...
  /api/profile:
    get:
      tags:
      - profile-api-controller
      operationId: getProfile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDto"
    put:
      tags:
      - profile-api-controller
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDto"
...
  /api/profile/change-password:
    post:
      tags:
      - profile-api-controller
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequestDto"
        required: true
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileApiChangePasswordErrors"
  /api/auth/token:
    post:
      tags:
      - authentication-api-controller
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/auth/login:
    post:
      tags:
      - authentication-api-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /api/auth/login-by-token:
    post:
      tags:
      - authentication-api-controller
      operationId: loginBySharedWorkspaceToken
      parameters:
      - name: sharedWorkspaceToken
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
components:
  schemas:
...
    I18nSettingsDto:
      required:
      - language
      - locale
      type: object
      properties:
        locale:
          type: string
        language:
          type: string
    UpdateProfileRequestDto:
      required:
      - i18n
      type: object
      properties:
        documentsStorage:
          maxLength: 255
          minLength: 0
          type: string
        i18n:
          $ref: "#/components/schemas/I18nSettingsDto"
    ProfileDto:
      required:
      - i18n
      - userName
      type: object
      properties:
        userName:
          type: string
        documentsStorage:
          type: string
        i18n:
          $ref: "#/components/schemas/I18nSettingsDto"
...
    ChangePasswordRequestDto:
      required:
      - currentPassword
      - newPassword
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    TokenResponse:
      required:
      - token
      type: object
      properties:
        token:
          type: string
...
    LoginRequest:
      required:
      - password
      - rememberMe
      - userName
      type: object
      properties:
        userName:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
...
    ProfileApiChangePasswordErrors:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          enum:
          - CurrentPasswordMismatch
          - TransientUser
          - NotAuthenticated
        message:
          type: string
...